requirements.txt
fastapi
uvicorn
python-multipart
openai
PyPDF2
python-docx
chromadb
sentence-transformers
requests
aiofiles